--- a/Marlin/dogm_bitmaps.h
+++ b/Marlin/dogm_bitmaps.h
@@ -373,3 +373,40 @@ const unsigned char extruder_graphic[EXTRUDERS][2][STATUS_EXTRUDER_BYTEWIDTH*STA
     }
   };
 #endif //HAS_FAN
+
+
+#define STATUS_SD0_WIDTH 8
+#define STATUS_SD0_BYTEWIDTH 1
+#define STATUS_SD0_HEIGHT 6
+
+#if ENABLED(SDSUPPORT)
+  const unsigned char sd0_graphic[STATUS_SD0_BYTEWIDTH*STATUS_SD0_HEIGHT] PROGMEM =
+  {
+     0xfc
+    ,0x82
+    ,0x81
+    ,0x81
+    ,0x81
+    ,0xff
+  };
+#endif //SDSUPPORT
+
+#define STATUS_SD1_WIDTH 8
+#define STATUS_SD1_BYTEWIDTH 1
+#define STATUS_SD1_HEIGHT 10
+
+#if ENABLED(SDSUPPORT)
+  const unsigned char sd1_graphic[STATUS_SD1_BYTEWIDTH*STATUS_SD1_HEIGHT] PROGMEM =
+  {
+    0xfc
+   ,0x82
+   ,0x81
+   ,0x81
+   ,0x81
+   ,0xff
+   ,0xff
+   ,0xff
+   ,0xff
+   ,0xff
+  };
+#endif //SDSUPPORT