--- a/Marlin/pins_RAMPS_13.h
+++ b/Marlin/pins_RAMPS_13.h
@@ -197,9 +197,7 @@
       // GLCD features
       //#define LCD_CONTRAST 190
       // Uncomment screen orientation
-      //#define LCD_SCREEN_ROT_90
       //#define LCD_SCREEN_ROT_180
-      //#define LCD_SCREEN_ROT_270
       //The encoder and click button
       #define BTN_EN1 40
       #define BTN_EN2 63