LCD-Speedup
A collection of little speed improvements, mainly for the graphical displays.
For a RAMPS, REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER, 1 etruder, bed, fan, SD-support the improvements sum up to ~ -33%. Expect more win with the other graphical displays.
For the character displays speed is about the same. Memory use has not changed significantly. See details below.

https://code.google.com/p/u8glib/wiki/tpictureloop helps to understand what is going on, when i write about stripes here.

![picture 165](https://cloud.githubusercontent.com/assets/211931/12279248/6df8604e-b986-11e5-88d1-8c3bfe813ae0.jpg)
![picture 167](https://cloud.githubusercontent.com/assets/211931/12279260/7b32b142-b986-11e5-976e-0b5a7c1def2b.jpg)
![picture 170](https://cloud.githubusercontent.com/assets/211931/12279265/81d7fae8-b986-11e5-8af5-15f924e13ebe.jpg)
Reduded perision if otherwise to wide:
![picture 171](https://cloud.githubusercontent.com/assets/211931/12279268/8547a494-b986-11e5-8e33-6e72e3a159ab.jpg)
![picture 182](https://cloud.githubusercontent.com/assets/211931/12279500/c74468e0-b987-11e5-9be8-43eee1db72eb.jpg)
![picture 172](https://cloud.githubusercontent.com/assets/211931/12279269/8c6caf6c-b986-11e5-8ee6-680b4b8dd03a.jpg)
![picture 173](https://cloud.githubusercontent.com/assets/211931/12279278/9668807c-b986-11e5-8721-fcaebe8ea9e5.jpg)
![picture 175](https://cloud.githubusercontent.com/assets/211931/12279308/b978cfb8-b986-11e5-9bc6-f9d260548340.jpg)
![picture 177](https://cloud.githubusercontent.com/assets/211931/12279322/c1cb3746-b986-11e5-9657-abdb926ec947.jpg)
![picture 178](https://cloud.githubusercontent.com/assets/211931/12279323/c427353a-b986-11e5-8e7f-369ac0ffe2ca.jpg)
![picture 180](https://cloud.githubusercontent.com/assets/211931/12279326/c854249c-b986-11e5-8e07-c8c462692b1b.jpg)
Update:
![picture 183](https://cloud.githubusercontent.com/assets/211931/12294105/102da4ac-b9f8-11e5-84e7-6ee973e37788.jpg)

---

Added a little benchmark.

Some tests to get the current state:

Tests with different display content CLCD
# define REPRAP_DISCOUNT_SMART_CONTROLLER

Der Sketch verwendet 75.344 Bytes (29%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 3.151 Bytes (38%) des dynamischen Speichers, 5.041 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 20-21ms
Homed 20-21ms
Extruder on 22ms
# define REPRAP_DISCOUNT_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1

Der Sketch verwendet 78.632 Bytes (30%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 3.165 Bytes (38%) des dynamischen Speichers, 5.027 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 23-24ms
Homed 23-24ms
Extruder on 24-25ms
Extruder + bed on 25ms
# define REPRAP_DISCOUNT_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 97.138 Bytes (38%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.392 Bytes (53%) des dynamischen Speichers, 3.800 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 25-26ms
Homed 26-27ms
Extruder on 26-27ms
Extruder + bed on 26-27ms
# define REPRAP_DISCOUNT_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

commented lcd_implementation_status_screen()
Der Sketch verwendet 95.138 Bytes (37%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.392 Bytes (53%) des dynamischen Speichers, 3.800 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 0ms
Homed 0ms
Extruder on 0ms
Extruder + bed on 0ms

Tests with different display content GLCD
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER

Der Sketch verwendet 84.372 Bytes (33%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 3.648 Bytes (44%) des dynamischen Speichers, 4.544 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 72ms
Homed 75-76ms
Fan100% 77-78ms
Extruder on 78-79ms
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1

Der Sketch verwendet 88.244 Bytes (34%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 3.662 Bytes (44%) des dynamischen Speichers, 4.530 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 72-73ms
Homed 76-77ms
Fan100% 78ms
Extruder on 79ms
Extruder + bed on 79ms
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.810 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 76-77ms
Homed 81ms
Fan100% 82ms
Extruder on 83ms
Extruder + bed on 84ms
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

commented lcd_implementation_status_screen()
Der Sketch verwendet 104.820 Bytes (41%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 29-30ms
Homed 29-30ms
Fan100% 29-30ms
Extruder on 29-30ms
Extruder + bed on 29-30ms
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER

  //U8GLIB_ST7920_128X64_RRD u8g(0);
  U8GLIB_ST7920_128X64_4X u8g(LCD_PINS_D4, LCD_PINS_ENABLE, LCD_PINS_RS); // original u8glib device
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

commented lcd_implementation_status_screen()
Der Sketch verwendet 105.460 Bytes (41%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.898 Bytes (59%) des dynamischen Speichers, 3.294 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 59-60ms
Homed 59-60ms
Fan100% 59-60ms
Extruder on 59-60ms
Extruder + bed on 59-60ms
// Marlins implementation of the device for the ST7920 is about 50% faster then the original u8glib device and 640bytes smaller!!!
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT
# define PAGE_HEIGHT 8   //128 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
commented lcd_implementation_status_screen()
Der Sketch verwendet 104.804 Bytes (41%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.505 Bytes (54%) des dynamischen Speichers, 3.687 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 29-30ms
Homed 29-30ms
Fan100% 29-30ms
Extruder on 29-30ms
Extruder + bed on 29-30ms
// With 8 instead of 2 stripes the empty update loop is not slower. Transfer graphic buffer to display. Saves 392bytes RAM (512-128=384).
# define U8G_10DELAY 3
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT
# define PAGE_HEIGHT 8   //128 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
commented lcd_implementation_status_screen()
Der Sketch verwendet 104.892 Bytes (41%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.505 Bytes (54%) des dynamischen Speichers, 3.687 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 27-28ms
Homed 27-28ms
Fan100% 27-28ms
Extruder on 27-28ms
Extruder + bed on 27-28ms
# define U8G_10DELAY 3
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

commented lcd_implementation_status_screen()
Der Sketch verwendet 104.908 Bytes (41%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 26-27ms
Homed 26-27ms
Fan100% 27-28ms
Extruder on 26-27ms
Extruder + bed on 26-27ms
# define U8G_10DELAY 3
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.898 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 73-74ms
Homed 78-79ms
Fan100% 79-80ms
Extruder on 81ms
Extruder + bed on 81ms
// With a reduced from 10 to 3 µs pause we can save about 2-3ms per display update. 2µs messed up my display - was to short. This may be different for other displays.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT
# define PAGE_HEIGHT 8  //128 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
Der Sketch verwendet 107.882 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.505 Bytes (54%) des dynamischen Speichers, 3.687 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 111-112ms
Homed 120-121ms
Fan100% 123ms
Extruder on 124-125ms
Extruder + bed on 125-126ms
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT
# define PAGE_HEIGHT 16   //256 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
Der Sketch verwendet 107.896 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.633 Bytes (56%) des dynamischen Speichers, 3.559 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 89ms
Homed 95-96ms
Fan100% 96-97ms
Extruder on 97ms
Extruder + bed on 98-99ms
// Increasing the number of stripes is disastrous. 7*stripes + 70 ~ time in ms
# define VIKI2

some changes to pins_RAMPS_13.h where needed to compile
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 103.650 Bytes (40%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.508 Bytes (55%) des dynamischen Speichers, 3.684 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 156-157ms
Homed 168-169ms
Fan100% 172-173ms
Extruder on 174-175ms
Extruder + bed on 175-176ms
// VIKI2 is known as extra slow. Here we see how slow. Thats more than 200% compared to ST7920VIKI2 and the other standard u8glib devises all use 8 stripes.
# define VIKI2

some changes to pins_RAMPS_13.h where needed to compile
  // U8GLIB_NHD_C12864 u8g(DOGLCD_CS, DOGLCD_A0); // 8 stripes
  U8GLIB_NHD_C12864_2X u8g(DOGLCD_CS, DOGLCD_A0); // 4 stripes
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 103.710 Bytes (40%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.636 Bytes (56%) des dynamischen Speichers, 3.556 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 125ms
Homed 133-134ms
Fan100% 135-136ms
Extruder on 136-137ms
Extruder + bed on 137ms
// Not good, but much better!!! 9.75*stripes + 98 ~ time in ms

Test again
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.898 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 76-77ms
Homed 81ms
Fan100% 82ms
Extruder on 83ms
Extruder + bed on 84ms
// unchanged

Remove the big heater graphic
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

//u8g.drawBitmapP(9,1,STATUS_SCREENBYTEWIDTH,STATUS_SCREENHEIGHT, (blink % 2) && fanSpeed ? status_screen0_bmp : status_screen1_bmp);
Der Sketch verwendet 107.272 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 67-68ms
Homed 71-72ms
Fan100% 73-74ms
Extruder on 73-74ms
Extruder + bed on 74-75ms

Remove the big heater graphic
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define SDSUPPORT

//u8g.drawBitmapP(9,1,STATUS_SCREENBYTEWIDTH,STATUS_SCREENHEIGHT, (blink % 2) && fanSpeed ? status_screen0_bmp : status_screen1_bmp);
Der Sketch verwendet 103.410 Bytes (40%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.875 Bytes (59%) des dynamischen Speichers, 3.317 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 67-68ms
Homed 71-72ms
Fan100% 73-74ms
Extruder on 73-74ms
// unchanged

Remove the big heater graphic
# define MOTHERBOARD BOARD_RAMPS_13_EEB
# define EXTRUDERS 1
# define TEMP_SENSOR_1 999
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

//u8g.drawBitmapP(9,1,STATUS_SCREENBYTEWIDTH,STATUS_SCREENHEIGHT, (blink % 2) && fanSpeed ? status_screen0_bmp : status_screen1_bmp);
Der Sketch verwendet 107.370 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 67-68ms
Homed 71-72ms
Fan100% 73-74ms
Extruder on 74-75ms
Extruder + bed on 74-75ms
All heaters on 74-75ms
# define MOTHERBOARD BOARD_RAMPS_13_EEB
# define EXTRUDERS 1
# define TEMP_SENSOR_1 999
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.996 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.889 Bytes (59%) des dynamischen Speichers, 3.303 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 76-77ms
Homed 80-81ms
Fan100% 81-82ms
Extruder on 83ms
Extruder + bed on 83-84ms
All heaters on 83-84ms
// drawing the big heater graphic needs about 10ms and is close to independent from the amount of drawn devices.

Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define MOTHERBOARD BOARD_RAMPS_13_EEB
# define EXTRUDERS 2
# define TEMP_SENSOR_1 999
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 112.280 Bytes (44%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.995 Bytes (60%) des dynamischen Speichers, 3.197 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 76-77ms
Homed 81-82ms
Fan100% 82-83ms
Extruder on 82-83ms
Extruder + bed on 83-84ms
All heaters on 83-84ms
// About the same speed

Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.650 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.891 Bytes (59%) des dynamischen Speichers, 3.301 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 72-73ms
Homed 77-78ms
Fan100% 78-79ms
Extruder on 79-80ms
Extruder + bed on 79-80ms

Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define SDSUPPORT

Der Sketch verwendet 103.418 Bytes (40%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.877 Bytes (59%) des dynamischen Speichers, 3.315 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 69-70ms
Homed 73-74ms
Fan100% 74-75ms
Extruder on 75-76ms

Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define SDSUPPORT
# define FAN_PIN           -1 // (Sprinter config)

Der Sketch verwendet 103.022 Bytes (40%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.875 Bytes (59%) des dynamischen Speichers, 3.317 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 65-66ms
Homed 70-71ms
Extruder on 70-71ms

Replace the inverse bar in the middle with boundary lines
Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.712 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.891 Bytes (59%) des dynamischen Speichers, 3.301 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
Fresh reset 60-61ms
Homed 64-65ms
Fan100% 65-66ms
Extruder on 66-67ms
Extruder + bed on 66-67ms
// About 5ms win!!!

Draw progress bar only when printing from SD / Alter SD symbol when SD-card is inserted.
Replace the inverse bar in the middle with boundary lines
Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 107.748 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.891 Bytes (59%) des dynamischen Speichers, 3.301 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
SD-card out
Fresh reset 55-56ms
Homed 59-60ms
Fan100% 60-61ms
Extruder on 61-62ms
Extruder + bed on 62-63ms
// About 5ms win!!!
SD-card in
Fresh reset 57-58ms
Homed 61-62ms
Fan100% 62-63ms
Extruder on 63-64ms
Extruder + bed on 63-64ms
// About 3ms win!!!
Printing from SD-card 
Extruder on 65-66ms
Extruder + bed on 65-66ms
// About 1ms win!!!

Pre-calculate strings. Update upper and lower half separate.
Draw progress bar only when printing from SD / Alter SD symbol when SD-card is inserted.
Replace the inverse bar in the middle with boundary lines
Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 108.058 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.944 Bytes (60%) des dynamischen Speichers, 3.248 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
SD-card out
Fresh reset 53-54ms
Homed 57-58ms
Fan100% 58-59ms
Extruder on 58-59ms
Extruder + bed on 59-60ms
// About 2ms win!!!
# define PAGE_HEIGHT 16  //256 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
Pre-calculate strings. Update upper and lower half separate.
Draw progress bar only when printing from SD / Alter SD symbol when SD-card is inserted.
Replace the inverse bar in the middle with boundary lines
Replaced the big heater graphic with small graphics for the single devices and draw only what is needed.
# define REPRAP_DISCOUNT_FULL_GRAPHIC_SMART_CONTROLLER
# define TEMP_SENSOR_BED 1
# define SDSUPPORT

Der Sketch verwendet 108.056 Bytes (42%) des Programmspeicherplatzes. Das Maximum sind 253.952 Bytes.
Globale Variablen verwenden 4.688 Bytes (57%) des dynamischen Speichers, 3.504 Bytes für lokale Variablen verbleiben. Das Maximum sind 8.192 Bytes.
SD-card out
Fresh reset 60-61ms
Homed 64-65ms
Fan100% 64-65ms
Extruder on 64-65ms
Extruder + bed on 65-66ms
//was before
Fresh reset 89ms
Homed 95-96ms
Fan100% 96-97ms
Extruder on 97ms
Extruder + bed on 98-99ms
// 4 stripes. 3*stripes + 54 ~ time in ms
# define PAGE_HEIGHT 8   //128 byte framebuffer

//#define PAGE_HEIGHT 32  //512 byte framebuffer
…