@tkurbad 2016-04-25
@AnHardt: I adapted your code to my [Deltaprintr branch of the current Marlin RC](https://github.com/tkurbad/Marlin/tree/RC_Deltaprintr). And what shall I say: The printer suddenly runs smoothly even at speeds beyond 80 mm/s.

Thank you very much for this extremely valueable improvement!