@AnHardt 2016-04-25
@tkurbad 
Thanks for your feedback.
May i ask what kind of display you are using?