@tkurbad 2016-04-25
@AnHardt 
It's a full graphic display with uSD slot and rotary encoder, very similar to The MaKrPanel.

With the default firmware (an early fork of Marlin with FSR-based bed leveling code by @jcrocholl) some blobbing occured while printing with the status screen on at speeds > 30 mm/s.

After adapting Johan's FSR bed leveling code to the newest Marlin RC, I had to realize that the steppers even stuttered during simple homing operations (not due to Johan's code but thanks to the more demanding new Marlin version) - which even made the FSR leveling impossible because of vibrations added by the stuttering.

After incorporating the changes of your branch and adding a [little hardware helper](https://github.com/JohnSL/FSR_Endstop) to simulate a FIX_MOUNTED_PROBE during FSR bed leveling, everything's running fine now and I'm finally satisfied with the overall performance of the printer.