@AnHardt 2016-04-25
Ah thanks. ST7565 was one of those i have not to test with. I just handled it analogous to other displays i had at hand.

Did you do the port commit by commit or all at once.
Especially i'd be interested in, if you could say, if the general speedup, or the last step, the better distribution of the display update gave the boost for your machine.