@tkurbad 2016-04-25
Unfortunately, I cannot tell, because I ported it all at once. But, judging from my previous experiences with the display, which becomes the more sluggish the bigger a single drawing operation becomes, I'd say that splitting up the bitmap in the upper half into smaller images and precalculating the strings alone would have given a huge boost in performance already. Anyway, I cannot say for sure...