@tkurbad 2016-04-26
@AnHardt 
Sorry for hijacking this PR.

I tried to merge the speedup into 1.1.0-RC6.
Everything seems to work, but I'm not quite sure if I did the conversion to the new LCDViewAction enum (which is used to instantiate lcdDrawUpdate in ultralcd.cpp) right - or better to say in a way that produces the most efficient LCD output. In several places things like `lcdDrawUpdate--` got replaced by switch-case statements.

If you find the time, can you please have a brief look at https://github.com/tkurbad/Marlin/tree/mergetest2 to see, if I got it at least marginally right?

Thank you very much in advance!