@AnHardt 2016-04-26
@tkurbad 
Sorry at the moment i have low to now time to look at this.
One day i'll understand what thinkyhead did and than will rework this PR.
When, one day, 1.1 is ready this will have a higher priority to me, since the is no chance for merging this into the main branch today, or the very near future.