@tkurbad 2016-04-27
@AnHardt 
I perfectly understand your reasoning.
Nonetheless if you don't mind, I'll try to keep your changes in sync with the progress of 1.1 RC for use in my fork of Marlin. So far it's running pretty smooth...