@AnHardt 2016-12-09
The most important ideas from here are now in Marlins RCBugFix.