@AnHardt 2016-12-09
I'll keep the branch here, to save the separated heater graphics.