@AnHardt 2016-12-15
You can now find the original graphics from here, plus a lot of additional suggestions in https://github.com/MarlinFirmware/Marlin/pull/5430#issuecomment-267124027 and following messages.